+++ src/module.ts
@@ -1,14 +1,9 @@
 import { AppPlugin } from '@grafana/data';
 
-import { App, Config } from './components';
+import { App } from './components';
 import { AppSettings } from './types';
 
 /**
  * App Plugin
  */
-export const plugin = new AppPlugin<AppSettings>().setRootPage(App).addConfigPage({
-  title: 'Configuration',
-  icon: 'cog',
-  body: Config,
-  id: 'configuration',
-});
+export const plugin = new AppPlugin<AppSettings>().setRootPage(App);