+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 
 - Add plugin e2e tests and remove cypress (#95)
 - Update to Grafana 11.1 dependencies (#97)
-- Update to Business Suite (#98)
+- Update to Business App (#98)
 - Update to Grafana 11.1.0 image (#99)
 
 ## 3.3.0 (2023-02-01)