+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Update to Grafana 11.1.4 image (#108)
+- Update to Business App (#109)
 
 ## 4.2.0 (2024-08-05)