+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features / Enhancements
 
 - Update to Grafana 11.1.4 image (#108)
+- Update to Business App (#109)
 
 ## 4.2.0 (2024-08-05)
 
@@ -29,7 +30,7 @@
 
 - Add plugin e2e tests and remove cypress (#95)
 - Update to Grafana 11.1 dependencies (#97)
-- Update to Business Suite (#98)
+- Update to Business App (#98)
 - Update to Grafana 11.1.0 image (#99)
 
 ## 3.3.0 (2023-02-01)