+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Features / Enhancements
 
-- Update to Grafana 11.1.4 image (#108)
-- Update to Business App (#109)
+- Updated to Grafana 11.1.4 image (#108)
+- Updated to Business App (#109)
+- Updated to Grafana 11.2.0 image with Scenes (#110)
 
 ## 4.2.0 (2024-08-05)